project - Copy/requirements.txt
Flask==3.0.3
torch>=2.0.1
torchvision>=0.15
Pillow==10.3.0
Werkzeug==3.0.1
# Optional if using EfficientNet-PyTorch package directly
efficientnet_pytorch==0.7.1
